feat: Add getScores(not complete)
