+++ src/main/java/com/bungaebowling/server/post/controller/PostController.java
@@ -123,26 +123,6 @@ public class PostController {
         return ResponseEntity.ok().body(response);
     }
 
-    @GetMapping("/{postId}/scores")
-    public ResponseEntity<?> getPostScores(@PathVariable Long postId) {
-        List<PostResponse.GetScoresDto.ScoreDto> scoreDtos = new ArrayList<>();
-        var getScoreDto1 = new PostResponse.GetScoresDto.ScoreDto(
-                1L,
-                180,
-                "/scoreImages/1.jpg"
-        );
-        scoreDtos.add(getScoreDto1);
-        var getScoreDto2 = new PostResponse.GetScoresDto.ScoreDto(
-                2L,
-                210,
-                null
-        );
-        scoreDtos.add(getScoreDto2);
-
-        var response = ApiUtils.success(scoreDtos);
-        return ResponseEntity.ok().body(response);
-    }
-
     // ToDo : 모집글 등록 response에 모집글 id 반환하기
     @PostMapping(consumes = {MediaType.APPLICATION_JSON_VALUE}) // json 타입만 처리 가능
     public ResponseEntity<?> createPost(
